requirements.txt
# Core Flask Dependencies (Required)
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
WTForms>=3.0.0

# Data Processing
pandas>=1.3.0
openpyxl>=3.0.0  # For Excel export

# Video Processing (for IMKH format support)
moviepy>=1.0.3
opencv-python>=4.8.0
imageio>=2.31.1
imageio-ffmpeg>=0.4.7
pillow>=9.0.0
numpy>=1.21.0

# AI Models for Person Detection
torch>=1.13.0
torchvision>=0.14.0
ultralytics>=8.0.0
onnxruntime>=1.15.0

# Transformer Models
transformers>=4.30.0
accelerate>=0.20.0
timm>=0.9.0

# SAM/SAM2 Models  
segment-anything>=1.0
sam2>=0.1.0

# Additional CV libraries
supervision>=0.20.0
roboflow>=1.0.0

# GPU Monitoring and Management
gputil>=1.4.0
nvidia-ml-py3>=7.352.0
psutil>=5.9.0

# Person Recognition (Machine Learning)
scikit-learn>=1.3.0
joblib>=1.3.0

# Face Recognition (Optional - can be difficult to install on Windows)
# Uncomment the following lines if you want face detection features:
# dlib>=19.24.0
# face_recognition>=1.3.0

# Optional Dependencies (install if available)
Flask-SocketIO>=5.0.0
Flask-Babel>=3.0.0
celery>=5.3.0
redis>=5.0.0

# OCR Text Extraction
easyocr>=1.7.0
pytesseract>=0.3.10

# Utilities
python-dotenv>=1.0.0
Werkzeug>=2.0.0
pandas>=2.0.0

# Production server (optional)
gunicorn>=20.0.0